backend/requirements.txt
fastapi==0.110.3
uvicorn==0.30.1
neo4j==5.21.0
pydantic==2.7.4
PyYAML==6.0.1
httpx==0.27.0

